+++ android/app/src/main/java/org/smartregister/fhircore/activity/PatientListActivity.kt
@@ -18,16 +18,10 @@ package org.smartregister.fhircore.activity
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import android.widget.Button
 import android.widget.EditText
-import android.widget.ImageButton
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.Toolbar
-import androidx.core.view.GravityCompat
 import androidx.core.widget.doAfterTextChanged
-import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
@@ -38,17 +32,16 @@ import org.smartregister.fhircore.util.Utils
 import org.smartregister.fhircore.util.Utils.addOnDrawableClickedListener
 
 /** An activity representing a list of Patients. */
-class PatientListActivity : AppCompatActivity() {
+class PatientListActivity : BaseSimpleActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    Log.d("PatientListActivity", "onCreate() called")
-    setContentView(R.layout.activity_patient_list)
-
-    val toolbar = findViewById<Toolbar>(R.id.toolbar)
-    setSupportActionBar(toolbar)
 
     setUpViews()
+  }
+
+  override fun getContentLayout(): Int {
+    return R.layout.activity_patient_list
   }
 
   private fun setUpViews() {
@@ -75,15 +68,6 @@ class PatientListActivity : AppCompatActivity() {
         )
         editText.addOnDrawableClickedListener(Utils.DrawablePosition.DRAWABLE_RIGHT) { it.clear() }
       }
-    }
-
-    setupDrawerContent()
-  }
-
-  private fun setupDrawerContent() {
-    val drawerLayout = findViewById<DrawerLayout>(R.id.drawer_layout)
-    findViewById<ImageButton>(R.id.btn_drawer_menu).setOnClickListener {
-      drawerLayout.openDrawer(GravityCompat.START)
     }
   }