+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Display message on empty list along with new client register button
+- Client count display in drawer
 
 ## [0.0.3] - 2021-05-24
 ### Added